requirements.txt
APScheduler==3.6.3
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
djangorestframework==3.11.1
idna==2.10
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tzlocal==2.1
urllib3==1.25.10
